split out configuration networking tests
